Restruture files
- Moves api functions from auth/api to uploads/api and shared/api
- Moves ui functions from auth/ui to uploads/ui and shared/ui
