+++ assets/scripts/auth/ui.js
@@ -1,11 +1,8 @@
 'use strict'
 
 const store = require('../store.js')
-const uploadsTableHandlebar = require('../templates/uploadsTable.handlebars')
-
-const clearForm = function (id) {
-  $(id)[0].reset()
-}
+const greenNotification = require('../shared/ui').greenNotification
+const redNotification = require('../shared/ui').redNotification
 
 const signInSuccess = function (data) {
   $('#sign-in-div').hide()
@@ -61,136 +58,6 @@ const signOutFailure = function () {
   signOutFailure()
 }
 
-const uploadFileSuccess = function () {
-  greenNotification('Uploaded file successfully')
-}
-
-const uploadFileFailure = function () {
-  redNotification('File failed to upload')
-}
-
-const viewFilesSuccess = function (files) {
-  $('#upload-table-container').empty()
-  $('#upload-table-container').append(uploadsTableHandlebar(files))
-}
-
-const viewFilesFailure = function () {
-  redNotification('Failed to get users files')
-}
-
-const deleteFileSuccess = function () {
-  store.uploadId = null
-  $('#confirmDeleteModal').modal('hide')
-  $('#view-file-div').hide()
-  $('#home-page').show()
-  greenNotification('File deleted')
-}
-
-const deleteFileFailure = function () {
-  $('#confirmDeleteModal').modal('hide')
-  redNotification('Failed to delete file')
-}
-
-const viewFileSuccess = function (response) {
-  store.uploadId = response.upload._id
-  greenNotification('File viewed')
-  $('#home-page').hide()
-  $('#view-file-div').show()
-  if (response.upload._owner !== store.user.id) {
-    $('#filename').prop('readonly', true)
-    $('#description').prop('readonly', true)
-    $('#tags_tag').prop('readonly', true)
-    $('#view-delete-button').hide()
-    $('#view-save-button').hide()
-  } else {
-    $('#filename').prop('readonly', false)
-    $('#description').prop('readonly', false)
-    $('#tags_tag').prop('readonly', false)
-    $('#view-delete-button').show()
-    $('#view-save-button').show()
-  }
-  $('#file-id').text(response.upload._id)
-  $('#filename').val(response.upload.filename)
-  $('#description').val(response.upload.description)
-  if (response.upload.tags) {
-    $('#tags').importTags(response.upload.tags)
-  }
-  $('#view-download-button').attr('href', response.upload._url)
-}
-
-const updateFileSuccess = function () {
-  store.uploadId = null
-  $('#view-file-div').hide()
-  $('#home-page').show()
-  greenNotification('File updated')
-}
-
-const updateFileFailure = function () {
-  store.uploadId = null // delete this line of code since if a user retries and the server is up again it won't work?
-  redNotification('Failed to update file')
-}
-
-const viewFileFailure = function () {
-  redNotification('Failed to view file')
-}
-
-const showDeleteModal = function () {
-  $('#confirmDeleteModal').modal('show')
-}
-
-const showHomePage = function () {
-  $('#view-file-div').hide()
-  $('#home-page').show()
-}
-
-const greenNotification = function (text, time = 1000, isDismissable = false) {
-  $.notify({
-    message: text
-  }, {
-    type: 'success',
-    placement: {
-      from: 'top',
-      align: 'center'
-    },
-    animate: {
-      enter: 'animated fadeInDown',
-      exit: 'animated fadeOutUp'
-    },
-    allow_dismiss: isDismissable,
-    z_index: 1100,
-    delay: time,
-    timer: 500,
-    template: '<div data-notify="container" class="col-11 col-sm-3 alert alert-{0}" role="alert">' +
-  '<button type="button" aria-hidden="true" class="close" data-notify="dismiss">×</button>' +
-  '<span data-notify="message">{2}</span>' +
-  '</div>'
-  })
-}
-
-const redNotification = function (text, time = 1000, isDismissable = false) {
-  $.notify({
-    message: text
-  }, {
-    type: 'danger',
-    placement: {
-      from: 'top',
-      align: 'center'
-    },
-    animate: {
-      enter: 'animated fadeInDown',
-      exit: 'animated fadeOutUp'
-    },
-    allow_dismiss: isDismissable,
-    z_index: 1100,
-    delay: time,
-    timer: 500,
-    template: '<div data-notify="container" class="col-11 col-sm-3 alert alert-{0}" role="alert">' +
-  '<button type="button" aria-hidden="true" class="close" data-notify="dismiss">×</button>' +
-  '<span data-notify="message">{2}</span>' +
-  '</div>'
-  })
-}
-
 module.exports = {
   signInSuccess,
   signInFailure,
@@ -201,18 +68,5 @@ module.exports = {
   changePasswordSuccess,
   changePasswordFailure,
   signOutSuccess,
-  signOutFailure,
-  uploadFileSuccess,
-  uploadFileFailure,
-  viewFilesSuccess,
-  viewFilesFailure,
-  deleteFileSuccess,
-  deleteFileFailure,
-  clearForm,
-  viewFileSuccess,
-  viewFileFailure,
-  showDeleteModal,
-  showHomePage,
-  updateFileSuccess,
-  updateFileFailure
+  signOutFailure
 }
