+++ assets/scripts/shared/api.js
@@ -1,8 +1,1 @@
 'use strict'
-
-const config = require('../config.js')
-const store = require('../store.js')
-
-module.exports = {
-
-}
