+++ assets/scripts/shared/events.js
@@ -1,18 +1,1 @@
 'use strict'
-
-const getFormFields = require(`../../../lib/get-form-fields`)
-
-const api = require('./api')
-const ui = require('./ui')
-const store = require('../store')
-
-
-
-const addHandlers = function () {
-
-}
-
-
-module.exports = {
-  addHandlers
-}
